fixed tests and refactored ChatItemGutter to ChatGutter
