+++ src/index.ts
@@ -21,6 +21,7 @@ export { default as ButtonGroup, ButtonGroupProps } from './components/Button/Bu
 
 export { default as Chat, ChatProps } from './components/Chat/Chat'
 export { default as ChatItem, ChatItemProps } from './components/Chat/ChatItem'
+export { default as ChatGutter, ChatGutterProps } from './components/Chat/ChatGutter'
 export { default as ChatMessage, ChatMessageProps } from './components/Chat/ChatMessage'
 
 export {
