+++ src/themes/teams/componentStyles.ts
@@ -11,7 +11,7 @@ export { default as ButtonGroup } from './components/Button/buttonGroupStyles'
 
 export { default as Chat } from './components/Chat/chatStyles'
 export { default as ChatItem } from './components/Chat/chatItemStyles'
-export { default as ChatItemGutter } from './components/Chat/chatItemGutterStyles'
+export { default as ChatGutter } from './components/Chat/chatGutterStyles'
 export { default as ChatMessage } from './components/Chat/chatMessageStyles'
 
 export { default as Divider } from './components/Divider/dividerStyles'
