+++ src/themes/teams/componentVariables.ts
@@ -9,7 +9,7 @@ export { default as ButtonGroup } from './components/Button/buttonVariables'
 
 export { default as Chat } from './components/Chat/chatVariables'
 export { default as ChatItem } from './components/Chat/chatItemVariables'
-export { default as ChatItemGutter } from './components/Chat/chatItemGutterVariables'
+export { default as ChatGutter } from './components/Chat/chatGutterVariables'
 export { default as ChatMessage } from './components/Chat/chatMessageVariables'
 
 export { default as Divider } from './components/Divider/dividerVariables'
