+++ src/themes/teams/components/Chat/chatGutterStyles.ts
@@ -1,0 +1,8 @@
+import { ICSSInJSStyle, ComponentSlotStylesInput } from '../../../types'
+import { ChatGutterProps } from '../../../../components/Chat/ChatGutter'
+
+const chatItemStyles: ComponentSlotStylesInput<ChatGutterProps> = {
+  root: (): ICSSInJSStyle => ({}),
+}
+
+export default chatItemStyles
