+++ test/specs/components/Chat/ChatGutter-test.ts
@@ -1,0 +1,6 @@
+import { isConformant } from 'test/specs/commonTests'
+import ChatGutter from 'src/components/Chat/ChatGutter'
+
+describe('ChatGutter', () => {
+  isConformant(ChatGutter)
+})
