+++ test/specs/components/Chat/ChatItem-test.tsx
@@ -1,5 +1,6 @@
 import { isConformant, implementsShorthandProp } from 'test/specs/commonTests'
 import ChatItem from 'src/components/Chat/ChatItem'
+import ChatGutter from 'src/components/Chat/ChatGutter'
 import Slot from 'src/components/Slot/Slot'
 
 const chatItemImplementsShorthandProp = implementsShorthandProp(ChatItem)
@@ -8,5 +9,5 @@ describe('ChatItem', () => {
   isConformant(ChatItem)
 
   chatItemImplementsShorthandProp('content', Slot, { mapsValueToProp: 'children' })
-  chatItemImplementsShorthandProp('gutter', Slot, { mapsValueToProp: 'children' })
+  chatItemImplementsShorthandProp('gutter', ChatGutter, { mapsValueToProp: 'content' })
 })
